+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ## HEAD (Unreleased)
 
-- (feature) Added iterator `Symbol.iterator`, and iterator methods `entries`, `values`, and `keys`
+- (breaking) Removed map function argument to `StridedView.random`, use `.map` or `.update` instead
+- (feature) Added iterator `Symbol.iterator`
+- (feature) and Array-like iterators methods `entries`, `values`, and `keys`
 - (feature) Added `inspect`
 - (feature) Added `StridedView.combine`
 - (feature) Static constructors now use `Float64Array` when possible
-- (feature) `StridedView.random` now generates random values in the range `[0, 1)` passed to the `fill` function
-- (feature) Added `StridedView.cwise`
+- (feature) Added `StridedView.zipWith`
 - (feature) Added `neighbors` iterator
 
 ---